backend/requirements.txt
pandas
mysql-connector-python
tqdm
pyarrow
flask
flask-cors
requests
